Fucking blindness bug, man
I've fixed that several times already, maybe now I managed to fix it properly?

+++ src/allmain.c
@@ -7478,7 +7478,7 @@ newbossX:
 
 		}
 
-		if (u.ucreamed < 0) {
+		if ((int)(u.ucreamed) < 0) {
 			pline("Fixing a bug that gave you a negative creamed counter...");
 			u.ucreamed = 0;
 		}
